requirements.txt
discord.py>=1.7.3
PyNaCl==1.4.0
async-timeout==3.0.1
ffmpeg==1.4
youtube-dl==2021.6.6
asyncpg==0.23.0